Don't use inline namespace "v2"
Even if SDL3 comes out soon, you're never going to be able to
link both SDL2 and SDL3 to the same programme, so versioning the
symbols isn't all that useful for that.

Where inline namespaces might be useful is if we decide at some
point that we want the bindings to be ABI stable, since they're
all inline. But we're a long way from that at the moment, so for
now just drop them.