Add experimental events wrapper using eggs::variant.
This, I suppose, is the "proper" way to do what SDL_Event does,
that is use a tagged union. However, while with SDL_Event you can
use a simple switch statement on the type() parameter, with
a variant you're "supposed" to use a visitor.

However, defining a visitor out-of-class is unpleasant, and we'd
much rather just be able to use lambda functions. So this uses a
trick I found on Stack Overflow to convert a list of lambda
functions to a visitor with multiple overloaded operator()s.

It all works (provided you include a generic lambda for the
default case), but... it's kind of hideously ugly...