Use a local copy of SDL2 if one is installed
This was what we did originally, dumped in favour of using a Git
submodule and building SDL that way.

But come to think of it, most users will probably be using an
installed copy, so let's test with that. It will (should) also
save CI build time.

Note that this approach still has SDL2 as a submodule and will
attempt to build and statically link it if a system installation
can't be found. This should allow future testing on iOS and other
platforms at some point.