Split utils.hpp into separate headers
There were so many clever metafunctions in there it was hard to
see what was going on. This commit splits it up into separate
sections.

Also, restrict the templated to_c_value() and from_c_value()
to only operate on fundamental types, enums and pointers, and
to take their arguments by value. It turns out the forwarding
reference versions were too greedy and were preventing some
customised versions begin called in WIP code. This also neatly
means we can't accidentelly forget to supply the functions for
more complex SDL types.

One thing that occurs to me is that allowing passthrough for any
pointer type is probably overly broad: it may be better to
restrict it only to pointer-to-fundemental types. I'll have a
think.