Add wrappers for SDL logging functions
This commit adds both simple wrappers for the log functions like

    SDL_Log(args...) -> sdl::log(args...)
    SDL_LogVerbose(args...) -> sdl::log(args...)
    etc

It also adds a more C++-y logger, inspired by Qt's log functions.
A bare call to

    sdl::log()

with no arguments will return an instance of the logger class. This
can be written to by any class which overloads operator<< for
ostreams, for example

    std::string msg1 = "Hello, ";
    std::string msg2 = "world";

    sdl::log() << msg1 << msg2;

The destructor of the logger will take care of calling the appropriate
SDL_Log() function.

Lastly, this commit provides a way to override the default logging
function, using a similar callback API used elsewhere in the library.