Add test for events overhead
I'm worried that using variants for events leads to overhead
compared with the plain C union.

It turns out it does: roughly 2-3% in release builds (whether
using sdl::variant::index() or a visitor), and a slightly
alarming 85% for debug builds.

I'm not sure whether this is a problem or not. After all, people
don't usually need to push a million messages a second through
the events system. Nonetheless, it's so fundemental to how SDL
works that I'm loathe to commit to something that doesn't vanish
pretty much completely in release builds.

And there's also the fact that sdl::variant won't even compile
with exceptions disabled...