[travis] Add XCode{6.4, 7.3} and GCC5 tests on OSX
Previously we were (rather arbitrarily) just testing on XCode 7.0.
This commit changes the OSX tests so that we use both the oldest
(6.4) and newest (7.3) Clang versions instead. It also adds testing
with g++ on OS X. Currently this is g++5.3, but I'd imagine Homebrew
is going to upgrade to 6.1 soon.