Rename window_view to window_ref
The "window_view" name was chosen for symmetry with the proposed
std::string_view class, as a non-owning reference to a window.

However, on reflection I think "window_ref" is probably a better
name, as in the context of a window, "view" sounds like it actually
displays something on-screen, when it does not. Furthermore, one
can change the properties of the original window via a
window_view/window_ref, so "view" might be misleading here.