[travis] Don't use the dummy video driver for OS X tests
The dummy video driver doesn't work very well on OS X, complaining
about missing OpenGL. The good news is we handle this correctly
(throwing an exception in the sdl::window constructor). The bad
news is that none of the tests work,

However, it seems like the normal video driver works on the
(presumably headless) Travis build servers for Mac OS, so let's
just use those.